requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
click==8.1.8
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
executing==2.2.0
fastapi==0.115.12
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Levenshtein==0.27.1
matplotlib-inline==0.1.7
motor==3.7.0
nest-asyncio==1.6.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
pytz==2025.2
pyzmq==26.4.0
RapidFuzz==3.13.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.1
wcwidth==0.2.13
